+++ processor/src/main/java/react4j/processor/ReactProcessor.java
@@ -145,7 +145,8 @@ public final class ReactProcessor
   {
     final String name = deriveComponentName( typeElement );
     final PackageElement packageElement = processingEnv.getElementUtils().getPackageOf( typeElement );
-    final ComponentDescriptor descriptor = new ComponentDescriptor( name, packageElement, typeElement );
+    final ComponentDescriptor descriptor =
+      new ComponentDescriptor( processingEnv.getSourceVersion(), name, packageElement, typeElement );
 
     determineComponentType( descriptor, typeElement );
     determineLifecycleMethods( typeElement, descriptor );